Downgrade Node version to 14 for deploy job
Node 15.6.0+ has a bug that causes serverless to produce a zipped
package with empty files. Deployment succeeds but the Lambda function
fails because handler.py is empty.

This is supposedly worked around in the latest serverless and its
dependencies, but upgrading to serverless v2+ causes other breaking
changes. Downgrade to Node 14 for now.

https://github.com/serverless/serverless/issues/8772