Fix fpm installation on openSUSE 15.4; switch to nfpm
fpm currently fails to install because the public_suffix gem dependency
requires Ruby 2.6, which openSUSE 15.4 doesn't provide. Because there have
been so many Ruby dependency issues with fpm in the past, switch to the fewer-
dependency nfpm, which also has better support for code signing if we want to
do that in the future.